// web/src/models/RemoteAnnouncement.ts
export type RemoteAnnouncement = {
    row_num: string;
    board_id: string;
    board_idx: string;
    board_subidx: number;
    board_subject: string;
    notice: string;
    member_type: string;
    member_idx: string;
    member_email: string;
    member_name: string;
    hit: number;
    use_tag: string;
    created_at: string;
};